move skipModelInput
